ButterQuant-with-deep-learning-1.5/backend/requirements.txt

Flask==3.0.3
flask-cors==4.0.1
yfinance==1.0
numpy==2.2.0
pandas==2.2.3
scipy==1.15.0
statsmodels==0.14.4
arch==7.2.0
packaging==24.2
redis==5.2.1
gevent==24.11.1
simplejson==3.19.3
PyYAML==6.0.2
psycopg2-binary==2.9.9
onnxruntime==1.20.1
joblib==1.5.3
ib_insync==0.9.86
nest_asyncio==1.6.0
scikit-learn==1.6.0
pyarrow>=14.0.0
torch>=2.0.0
onnxruntime-gpu>=1.15.0  # Optional for GPU support
matplotlib>=3.8.0
seaborn>=0.13.0
onnx>=1.15.0
onnxscript>=0.1.0
alpaca-py>=0.37.0
python-dotenv>=1.0.1
